Created deep thoughts api and controller

+++ database/migrations/2025_02_21_163945_add_user_id_to_blogs_table.php
@@ -7,9 +7,11 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up()
     {
-        Schema::table('blogs', function (Blueprint $table) {
-            $table->foreignId('user_id')->constrained()->onDelete('cascade'); // Add user_id column
-        });
+        if (!Schema::hasColumn('blogs', 'user_id')) {
+            Schema::table('blogs', function (Blueprint $table) {
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+            });
+        }
     }
 
     public function down()
